URBNG-35 Initial version of WebAPI implementation
This commit includes:
+ authentication - simple HTTP Basic Authentication
+ get list of games command
+ get game details (dynamic and static)
+ get user games

It's working on mock with hardcoded database.
Responses are not in Hal+Json format yet, only Json or string messages
and request arguments are ignored in this initial version.

Tested only by hand, unit test will be written as soon as I implement
Hal+Json format response.